Rewrote back-ends, incorporating the code for extracting structs from query results.
Added partially complete direct-find functionality, which is a fast way of getting/caching structs by guid.

The back-ends are also rewritten using objects rather than units, which makes it easier to overwrite
and customise bits of code, and access the Snooze object from the back-end. Access to the Snooze object
will make it simpler to cache structs when processing query results.