Reimplement the connection pool.
Reimplement the connection pool using a simpler design. The maximum number of connections is specified at startup and connections are reused. Application threads will block till a connection is available if they are all in use at the time. Connections are managed in a kill-safe manner: if a application thread dies before returning a connection that case will be caught and the connection recovered.