Implement proportional error control for dynamically resizing the connection pool. PE control works by applying a control signal (in our case, creating or freeing connections) based on the difference (or error) between a target and actual value. In the case of the connection pool this means:
- We measure the demand for connections. This is the sum of
  the number of connections in use and the number of threads
  waiting for a connection

- We record how many connections exist in total (those in use
  and those in the pool)

- The difference is the error.

- We add (or remove) gain * error connections. (We clip this
  value to make sure we don't go outside min- and max-connections

PE control works quite well, but there is a lag. If this lag
is appreciable we can add additional terms (i.e. PID control).
I have also added sufficient logging to reproduce the history
of messages to the connection pool, so we can gain greater
insight into how it is working.