Fixed issue in postgresql8 back-end.
If a client thread dies at an inopportune moment, it's possible for an spgsql connection to be left in a state where all calls to query methods block indefinitely.

There is a semaphore in spgsql that's used to make sure only one query method is run at a time. It's sometimes possible for this semaphore to be left in a non-zero state. All subsequent queries block, including ROLLBACK queries and graceful disconnects (which send a "Terminate" message that is also blocked).

The reset-connection method in postgresql8-database% sends "ROLLBACK" to allow the connection pool to recycle a connection after a thread death. If reset-connection blocks, the connection manager also blocks, and the connection pool stops working.

The fix reimplements reset-connection. Rather than sending "ROLLBACK", it actually impolitely disconnects and reconnects the spgsql back-end to the connection. Impolite disconnects amount to simply closing the TCP/IP port, so they don't block.